refactor(YB-148): 도메인 모듈에서 공통으로 사용되는 enum 들을 common/enums 에 위치
