+++ yellobook-domain/src/main/java/com/yellobook/common/enums/OrderStatus.java
@@ -1,0 +1,21 @@
+package com.yellobook.common.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum OrderStatus {
+    /**
+     * 주문 확인 대기중 (DEFAULT)
+     */
+    PENDING_CONFIRM,
+    /**
+     * 주문 정정 대기중
+     */
+    PENDING_MODIFY,
+    /**
+     * 주문 확인 완료
+     */
+    CONFIRMED,
+}
