+++ yellobook-domain/src/main/java/com/yellobook/common/enums/ScheduleType.java
@@ -1,0 +1,12 @@
+package com.yellobook.common.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum ScheduleType {
+    INFORM("INFORM"),
+    ORDER("ORDER");
+    private final String value;
+}
