--- yellobook-domain/src/main/java/com/yellobook/enums/OrderStatus.java
@@ -1,12 +1,0 @@
-package com.yellobook.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum OrderStatus {
-    PENDING_CONFIRM,
-    PENDING_MODIFY,
-    CONFIRMED,
-}
